speed improvements
change to a vector from unordered_map for NoAllocLogRecord.  Also add a NopLogExporter for testing.  Change add_attribute to take a const char* so strlen does not get called.  Also correct calculation of CirculeBuffer::size.  Was allowing multiple entries to be added/removed while size was calculated causing it to return negative values